refactor(react-gui): give the render pipeline a folder and its seams
`renderJob.service.ts` was 1,401 lines holding four render pipelines at
once -- (still | movie) x (external process | in-process thread) -- with a
shared registry, a shared poll timer and a shared progress model threaded
through all of them. Reading any one of the four meant stepping over the
other three.

`renderjob/` now has a file per stage of a job's life. `types.ts` says what
a job is made of; `jobRegistry.ts` holds the in-flight ones and the teardown
every exit path shares (the timer and the animation, which have to stop
whether the job completed, failed or was cancelled); `startJob.ts` has the
three starters that are the branches of `renderStart`; `jobLoop.ts` is one
timer tick working out which phase to poll; `completion.ts` is where a
finished unit goes; `inProcessJob.ts` and `movieJob.ts` are the two machines
with state of their own. `renderBackends/` becomes `renderjob/backends/`.

Where the seams are was decided by the call graph, not by subject. Splitting
the movie's frames from its encode reads well and does not work: rendering a
frame leads either to the next frame or to the encode, and the encode ends
at the same finish the last frame would have reached, so the two halves
reference each other. They are one state machine and they are one file. What
does come out is `encodeSpec.ts` -- is there an encode, where is ffmpeg,
what options -- which is questions with answers rather than transitions, and
depends on nothing above it.

The registry glob widens to one level of folders, which makes a name
registered twice reachable in a way it was not before: a folder barrel next
to the flat services it was split out of. `WorkerService.register` only warns
and lets the later one win, so `index.test.ts` now pins that each name is
registered exactly once, that every registered name is one the call contract
declares, and that a folder's services are found at all.
