+++ tritium/react-gui/src/renderer/__test__/ffmpegEncode.test.ts
@@ -11,7 +11,7 @@ import {
   movieFileName,
   movieOutputPath,
   type FfmpegEncodeOptions,
-} from '@renderer/worker/server/services/ffmpegEncode';
+} from '@renderer/worker/server/services/renderjob/ffmpegEncode';
 
 const base: FfmpegEncodeOptions = {
   outputDir: '/out',
