+++ tritium/react-gui/src/renderer/__test__/umbreonBackend.test.ts
@@ -15,7 +15,7 @@ import { describe, it, expect, vi } from 'vitest'
 import {
     umbreonBackend,
     umbreonNprBackend,
-} from '@renderer/worker/server/services/renderBackends/UmbreonBackend'
+} from '@renderer/worker/server/services/renderjob/backends/UmbreonBackend'
 import type { WorkerContext } from '@renderer/worker/server/types/WorkerContext'
 import type { RenderSettingsSnapshot } from '@renderer/data/renderResult'
 import type { PropDef } from '@renderer/data/rendererProperties'
