+++ tritium/react-gui/src/renderer/worker/server/WorkerService.ts
@@ -35,7 +35,7 @@ import type {
     ServiceKey,
 } from '@renderer/worker/shared/calls';
 import { forgetAnimProgress, pauseInactivePlayback, pumpAnimProgress } from '@renderer/worker/server/services/animation.service';
-import { isSceneBeingRendered } from '@renderer/worker/server/services/renderJob.service';
+import { isSceneBeingRendered } from '@renderer/worker/server/services/renderjob/renderJob.service';
 
 // import { createLogger } from '@cuemol/core/src/logger';
 // const log = createLogger(import.meta.url);
